backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pillow>=10.4.0
requests>=2.32.3
python-dotenv>=1.0.1
transformers==4.43.3
mysql-connector-python>=9.0.0
numpy<2.0
python-multipart
torch==2.3.1
torchvision==0.18.1
